Update Solana to 1.16.1 and Rust to 1.69 (#4592)
#### Problem

The 1.16 Solana crates are out, but SPL is still on 1.14 and needs the
new functionality.

#### Solution

Update the:

* rust version to 1.69
* nightly to 2023-04-19
* solana tools to 1.16.1
* solana crates to 1.16.1
* borsh to 0.10

And fix:

* new clippy warnings
* deprecated warnings in the new solana crates

+++ token-swap/program/src/curve/offset.rs
@@ -474,9 +474,9 @@ mod tests {
                 (swap_source_amount * swap_destination_amount));
             check_curve_value_from_swap(
                 &curve,
-                source_token_amount as u128,
-                swap_source_amount as u128,
-                swap_destination_amount as u128,
+                source_token_amount,
+                swap_source_amount,
+                swap_destination_amount,
                 TradeDirection::AtoB
             );
         }
@@ -502,9 +502,9 @@ mod tests {
             prop_assume!(!(swap_source_amount + token_b_offset).overflowing_mul(swap_destination_amount).1);
             check_curve_value_from_swap(
                 &curve,
-                source_token_amount as u128,
-                swap_source_amount as u128,
-                swap_destination_amount as u128,
+                source_token_amount,
+                swap_source_amount,
+                swap_destination_amount,
                 TradeDirection::BtoA
             );
         }
